move CUBEfunction to analytic functions
