+++ src/analyticfunctions/CMakeLists.txt
@@ -2,4 +2,5 @@ target_sources(mrchem PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/HydrogenFunction.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/NuclearFunction.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/NuclearGradientFunction.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/CUBEfunction.cpp
     )
