+++ src/analyticfunctions/CUBEfunction.cpp
