+++ src/analyticfunctions/CUBEfunction.h
