+++ src/qmfunctions/CMakeLists.txt
@@ -6,5 +6,4 @@ target_sources(mrchem PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/qmfunction_utils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/orbital_utils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/density_utils.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/CUBEfunction.cpp
     )
